Change the type of variable resto from int to  float

+++ src/brushless.h
@@ -12,7 +12,7 @@ typedef struct _startup_data{
    int min;
    float dec;
    int currentValue;
-   int resto;
+   float resto;
 } *startupData;
